Fix bug in standalone where it could not start f there was no database_path specified on command line and no previous database recorded
Now it does the correct sequence; if no database on command line and no previous database
recorded, puts up a message explaining situation and then open database window.

Change AbstractDBContainer.readEnvironment to not throw but return null if database not
selected; callers were already expecting it to return null in that circumstance.  Amended
javadoc to make clear that it would return null.